+++ pyproject.toml
@@ -6,6 +6,9 @@ extend-exclude = ["build", "dist"]
 [tool.ruff.lint]
 select = ["E4", "E7", "E9", "F", "I", "B", "UP"]
 
+[tool.ruff.lint.isort]
+known-first-party = ["tiktodv3", "verification"]
+
 [tool.mypy]
 python_version = "3.10"
 check_untyped_defs = true
