+++ tiktodv3/__init__.py
@@ -1,0 +1,7 @@
+"""TIKTOD V3 desktop application package."""
+
+from .config import APP_NAME, APP_VERSION
+
+__version__ = APP_VERSION
+
+__all__ = ["APP_NAME", "APP_VERSION", "__version__"]
