+++ tiktodv3/__main__.py
@@ -1,0 +1,6 @@
+"""Run TIKTOD V3 with ``python -m tiktodv3``."""
+
+from .gui import main
+
+if __name__ == "__main__":
+    main()
