+++ tiktodv3/settings.py
@@ -1,3 +1,5 @@
+"""Persistent user preference storage."""
+
 import json
 from dataclasses import asdict, dataclass
 from pathlib import Path
