+++ tiktodv3/utils.py
@@ -3,7 +3,7 @@ import sys
 import time
 from pathlib import Path
 
-PROJECT_ROOT = Path(__file__).resolve().parent
+PROJECT_ROOT = Path(__file__).resolve().parent.parent
 
 
 def log_message(app, message: str) -> None:
